:ambulance: fix: コンポーネント分けを修正
